@@ -20,12 +20,17 @@ SQLAlchemy = "^2.0.0"
 dynaconf = "^3.1.12"
 geopandas = "^0.14.1"
 ipython = "^8.5.0"
-"macrostrat.app-frame" = "^1.2.4"
-"macrostrat.database" = "^3.1.1"
+"macrostrat.app-frame" = "^2.0.0"
+"macrostrat.database" = "^3.3.2"
+"macrostrat.utils" = "^1.2.2"
+"macrostrat.dinosaur" = "^3.0.1"
+"macrostrat.core" = { path = "../core", develop = true }
+"macrostrat.map-integration" = { path = "../map-integration", develop = true }
+"criticalmaas.ta1-geopackage" = "^0.2.0"
 numpy = "^1.23.4"
 psycopg2-binary = "^2.9.4"
 pyproj = "^3.4.0"
-python = "^3.10"
+python = "^3.11"
 python-dotenv = "^1.0.0"
 requests = "^2.28.1"
 rich = "^13"
@@ -34,19 +39,19 @@ shapely = "^2.0.1"
 spectra = "^0.0.11"
 tiletanic = "^1.1.0"
 tqdm = "^4.65.0"
-typer = "^0.9.0"
+typer = "^0.12"
 aiofiles = "^23.2.1"
-docker = "^6.1.3"
+docker = "^7"
 toml = "^0.10.2"
-criticalmaas-ta1-geopackage = "^0.2.0"
 greenlet = "^3.0.3"
-macrostrat-dinosaur = "^3.0.1"
+paramiko = "^3.4.0"
 
 [tool.poetry.scripts]
 macrostrat = "macrostrat.cli.cli:main"
 
 [tool.poetry.group.dev.dependencies]
 snakeviz = "^2.2.0"
+"macrostrat.package_tools" = "^1.1.0"
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
