@@ -1,6 +1,8 @@
-from ..base import Migration, exists, has_fks
-from macrostrat.database import Database
 from pathlib import Path
+
+from macrostrat.database import Database
+
+from ..base import Migration, exists, has_fks
 
 __dir__ = Path(__file__).parent
 
@@ -12,6 +14,8 @@ class MacrostratCoreMigration(Migration):
     description = """
     Populate the `macrostrat` schema of the database prior to importing data from mariadb
     """
-    depends_on = ['baseline']
+    depends_on = ["baseline"]
 
-    postconditions = [exists('macrostrat', 'projects', 'sections', 'strat_tree', 'unit_boundaries')]
+    postconditions = [
+        exists("macrostrat", "projects", "sections", "strat_tree", "unit_boundaries")
+    ]
