@@ -1,13 +1,15 @@
-from macrostrat.database import Database
-from ..base import Migration, ApplicationStatus
-
 from psycopg2.sql import Identifier
+
+from macrostrat.database import Database
+
+from ..base import ApplicationStatus, Migration
 
 MATCHES_SLUG_SQL = """
 SELECT table_name
 FROM information_schema.tables
 JOIN maps.sources ON slug = table_name
 WHERE table_schema = 'sources'"""
+
 
 class MapSourceSlugsMigration(Migration):
     name = "map-source-slug"
@@ -17,7 +19,7 @@ class MapSourceSlugsMigration(Migration):
     then add a '_polygons' suffix to the associated primary table name in the sources schema.
     """
 
-    depends_on = ['baseline']
+    depends_on = ["baseline"]
 
     destructive = True
 
@@ -27,9 +29,9 @@ class MapSourceSlugsMigration(Migration):
 
         # Then, manually rename all primary_tables
         self._rename_primary_tables(db)
-    
+
     def _rename_primary_tables(self, db: Database):
-        for table_name, in db.run_query(MATCHES_SLUG_SQL):
+        for (table_name,) in db.run_query(MATCHES_SLUG_SQL):
             new_table_name = table_name + "_polygons"
             db.run_sql(
                 "ALTER TABLE sources.{table_name} RENAME TO {new_table_name}",
@@ -38,21 +40,22 @@ class MapSourceSlugsMigration(Migration):
                     new_table_name=Identifier(new_table_name),
                 ),
             )
-    
+
     def should_apply(self, db: Database):
         insp = db.inspector
 
         # Check that maps.sources exists, and has a 'slug' column
-        if not insp.has_table('sources', 'maps'):
+        if not insp.has_table("sources", "maps"):
             return ApplicationStatus.CANT_APPLY
 
-        col_names = [c['name'] for c in insp.get_columns('sources','maps')]
-        if not 'slug' in col_names:
+        col_names = [c["name"] for c in insp.get_columns("sources", "maps")]
+        if not "slug" in col_names:
             return ApplicationStatus.CAN_APPLY
 
         # Check that the primary_table column has appropriate values
         non_polygon_table_names = db.run_query(
-            "SELECT primary_table FROM maps.sources WHERE primary_table NOT LIKE '%_polygons' AND primary_table != 'unknown'")
+            "SELECT primary_table FROM maps.sources WHERE primary_table NOT LIKE '%_polygons' AND primary_table != 'unknown'"
+        )
         if non_polygon_table_names.first() is not None:
             return ApplicationStatus.CAN_APPLY
 
@@ -62,4 +65,3 @@ class MapSourceSlugsMigration(Migration):
             return ApplicationStatus.CAN_APPLY
 
         return ApplicationStatus.APPLIED
-
