@@ -1,6 +1,8 @@
-from ..base import Migration, ApplicationStatus, exists, view_exists
-from macrostrat.database import Database
 from pathlib import Path
+
+from macrostrat.database import Database
+
+from ..base import ApplicationStatus, Migration, exists, view_exists
 
 __dir__ = Path(__file__).parent
 
@@ -13,7 +15,7 @@ class PartitionMapsMigration(Migration):
     Starting from a Macrostrat v1 map database (burwell), integrate the tiny, small, medium, and large map tables (+lines)
     into a single map table, partitioned by scale to retain the original table structure and avoid copying data.
     """
-    depends_on = ['macrostrat-core-v2']
+    depends_on = ["macrostrat-core-v2"]
 
     postconditions = [
         exists("maps", "lines", "polygons"),
