@@ -2,9 +2,11 @@
 # https://github.com/UW-Macrostrat/postgis-tile-utils @ 1ef18f99436eff1d833b6b0463946644f249763a
 # https://github.com/UW-Macrostrat/tileserver @ v2.1.0
 
-from ..base import Migration, schema_exists
-from macrostrat.database import Database
 from pathlib import Path
+
+from macrostrat.database import Database
+
+from ..base import Migration, schema_exists
 
 __dir__ = Path(__file__).parent
 
@@ -15,6 +17,6 @@ class TileserverMigration(Migration):
     description = """
     Populate the `tile_cache` and `tile_utils` schemas
     """
-    depends_on = ['macrostrat-mariadb']
+    depends_on = ["macrostrat-mariadb"]
 
-    postconditions = [schema_exists('tile_cache'), schema_exists('tile_utils')]
+    postconditions = [schema_exists("tile_cache"), schema_exists("tile_utils")]
