@@ -1,6 +1,8 @@
-from ..base import Migration, exists, has_fks
-from macrostrat.database import Database
 from pathlib import Path
+
+from macrostrat.database import Database
+
+from ..base import Migration, exists, has_fks
 
 __dir__ = Path(__file__).parent
 
@@ -14,9 +16,9 @@ class MacrostratCoreMigration(Migration):
     Update the Macrostrat core schema ('macrostrat') with foreign keys and data transformations to
     stabilize the schema in PostgreSQL after transformation from MariaDB.
     """
-    depends_on = ['macrostrat-mariadb', 'api-v3']
+    depends_on = ["macrostrat-mariadb", "api-v3"]
 
     postconditions = [
-        exists('macrostrat', 'units', 'sections'),
-        has_fks('macrostrat', 'units', 'sections')
+        exists("macrostrat", "units", "sections"),
+        has_fks("macrostrat", "units", "sections"),
     ]
