@@ -1,3 +1,5 @@
+from typer import Typer
+
 from macrostrat.core import MacrostratSubsystem
 
 
@@ -5,8 +7,10 @@ class MapboardSubsystem(MacrostratSubsystem):
     name = "mapboard"
 
     def on_add_commands(self, cli):
-        @cli.command("export-mapboard", rich_help_panel="Export")
+        cmd = Typer(no_args_is_help=True, short_help="Macrostrat Mapboard commands")
+
+        @cmd.command("export-mapboard", rich_help_panel="Export")
         def export_map():
             print("Hello, world!")
 
-        cli.add_typer(self.control_command(), rich_help_panel="Subsystems")
+        cli.add_typer(cmd, name="mapboard", rich_help_panel="Subsystems")
