@@ -3,13 +3,10 @@ from pathlib import Path
 
 from dynaconf import Dynaconf, Validator
 from sqlalchemy.engine import make_url
+from sqlalchemy.engine.url import URL
 from toml import load as load_toml
 
 from .utils import find_macrostrat_config, is_pg_url
-
-cfg = find_macrostrat_config()
-
-macrostrat_config_file = cfg
 
 
 class MacrostratConfig(Dynaconf):
@@ -17,14 +14,21 @@ class MacrostratConfig(Dynaconf):
 
     def __init__(self, *args, **kwargs):
         cfg = find_macrostrat_config()
+        settings = []
+        if cfg is not None:
+            settings.append(cfg)
+
         super().__init__(
             envvar_prefix="MACROSTRAT",
             environments=True,
             env_switcher="MACROSTRAT_ENV",
-            settings_files=[cfg],
+            settings_files=settings,
             load_dotenv=False,
         )
-        self.config_file = Path(cfg)
+
+        self.config_file = None
+        if cfg is not None:
+            self.config_file = Path(cfg)
 
     def all_environments(self):
         # Parse out top-level headers from TOML file
@@ -53,6 +57,12 @@ PG_DATABASE = settings.pg_database
 # On mac and windows, we need to use the docker host `host.docker.internal` or `host.lima.internal`, etc.
 docker_localhost = getattr(settings, "docker_localhost", "localhost")
 PG_DATABASE_DOCKER = PG_DATABASE.replace("localhost", docker_localhost)
+
+mysql_database = getattr(settings, "mysql_database", None)
+if mysql_database is not None:
+    mysql_database: URL = make_url(mysql_database).set(drivername="mysql+pymysql")
+    # TODO: handle this more intelligently
+
 
 if elevation_database := getattr(settings, "elevation_database", None):
     environ["ELEVATION_DATABASE_URL"] = elevation_database
