@@ -3,12 +3,14 @@ from pathlib import Path
 
 import toml
 from dynaconf import Dynaconf
-from macrostrat.app_frame import Application, Subsystem, SubsystemManager
-from macrostrat.utils import get_logger
 from rich import print
 from rich.console import Console
-from typer import Typer, get_app_dir, Context
+from typer import Context, Typer, get_app_dir
 from typer.core import TyperGroup
+
+from macrostrat.app_frame import Application, Subsystem, SubsystemManager
+from macrostrat.app_frame.control_command import OrderCommands
+from macrostrat.utils import get_logger
 
 from .console import console_theme
 from .exc import MacrostratError
@@ -64,12 +66,6 @@ def load_settings():
         )
 
     return settings
-
-
-class OrderCommands(TyperGroup):
-    def list_commands(self, ctx: Context):
-        """Return list of commands in the order appear."""
-        return list(self.commands)  # get commands using self.commands
 
 
 class MacrostratSubsystem(Subsystem):
