@@ -1,9 +1,13 @@
-import sys, os, time
-from subprocess import call
 import argparse
+import os
+import sys
+import time
+from subprocess import call
+
 import psycopg2
-from psycopg2.extensions import AsIs
 import yaml
+from psycopg2.extensions import AsIs
+
 with open('../credentials.yml', 'r') as f:
     credentials = yaml.load(f)
 
