@@ -1,5 +1,6 @@
-from dotenv import load_dotenv
 from os import getenv
+
+from dotenv import load_dotenv
 
 load_dotenv()
 
