@@ -1,11 +1,13 @@
 import pytest
-from .fixtures import get_sql, Database
+
+from .fixtures import Database, get_sql
 
 
 @pytest.fixture(scope="session")
 def db_():
     db = Database()
     return db
+
 
 @pytest.fixture(scope="session")
 def setup(db_):
@@ -16,6 +18,7 @@ def setup(db_):
         cur.execute("DROP SCHEMA IF EXISTS macrostrat CASCADE;")
         cur.execute(schema)
         cur.execute(data_inserts)
+
 
 @pytest.fixture()
 def db(setup, db_):
